fix: widen the CUDA stream parameter to intptr_t and forward it in the distance launchers (#25)
A `cudaStream_t` is a pointer, hence 64-bit on every platform we build for,
but the public `stream` parameter was declared `int` throughout the CUDA
launchers. Any real (non-null) stream handle with a non-zero upper half is
therefore truncated on the way in and reinterpreted as a different, invalid
stream -- silent corruption rather than a diagnosable error. Widen every
launcher's `stream` parameter to `intptr_t` so the handle survives the trip.

The euclidean, L1 and mesh launchers additionally accepted no stream at all:
they hardcoded the default stream in the launch configuration
(`<<<blocks, threads, 0>>>` leaves the stream argument off) and uploaded
shape/stride metadata with the blocking `cudaMemcpy`, so every call also
served as a synchronisation point. Thread the stream through `dt()`/`sdt()`
and launch on it, mirroring the posdef/resize/restrict/splinc launchers that
already forwarded it.

`copyToDeviceAsync` is added alongside `copyToDevice` in `utils.h` for those
metadata uploads: it enqueues `cudaMemcpyAsync` on the caller's stream. The
converting path synchronises the stream before releasing its pinned staging
buffer, which the DMA engine reads asynchronously; the same-type path needs
no wait because `cudaMemcpyAsync` from pageable memory does not return until
the source has been staged. The existing synchronous `copyToDevice` is left
untouched so no other module changes behaviour.

Verified by compiling fastfields-cuda-lib's `distance.o` and `posdef.o`
against these headers with nvcc 12.0. There is no GPU available, so the
runtime behaviour of the stream plumbing is unverified by construction.

Refs fastfields/fastfields-lib#4

+++ include/fastfields/impl/cuda/posdef.h
@@ -319,7 +319,7 @@ void sym_matvec(
     const offset_t * stride_out,
     const offset_t * stride_hes,
     const offset_t * stride_inp,
-          int        stream = 0)
+          intptr_t   stream = 0)
 {
     const offset_t ndim = nbatch + 1;
     offset_t * d_size = nullptr, * d_so = nullptr, * d_sh = nullptr, * d_si = nullptr;
@@ -351,7 +351,7 @@ void sym_matvec_backward(
     const offset_t * stride_out,
     const offset_t * stride_grd,
     const offset_t * stride_inp,
-          int        stream = 0)
+          intptr_t   stream = 0)
 {
     const offset_t ndim = nbatch + 1;
     offset_t * d_size = nullptr, * d_so = nullptr, * d_sg = nullptr, * d_si = nullptr;
@@ -383,7 +383,7 @@ void sym_addmatvec_(
     const offset_t * stride_out,
     const offset_t * stride_hes,
     const offset_t * stride_inp,
-          int        stream = 0)
+          intptr_t   stream = 0)
 {
     const offset_t ndim = nbatch + 1;
     offset_t * d_size = nullptr, * d_so = nullptr, * d_sh = nullptr, * d_si = nullptr;
@@ -415,7 +415,7 @@ void sym_submatvec_(
     const offset_t * stride_out,
     const offset_t * stride_hes,
     const offset_t * stride_inp,
-          int        stream = 0)
+          intptr_t   stream = 0)
 {
     const offset_t ndim = nbatch + 1;
     offset_t * d_size = nullptr, * d_so = nullptr, * d_sh = nullptr, * d_si = nullptr;
@@ -449,7 +449,7 @@ void sym_solve(
     const offset_t * stride_inp,
     const offset_t * stride_hes,
     const offset_t * stride_wgt,
-          int        stream = 0)
+          intptr_t   stream = 0)
 {
     const offset_t ndim = nbatch + 1;
     offset_t * d_size = nullptr, * d_so = nullptr, * d_si = nullptr,
@@ -483,7 +483,7 @@ void sym_solve_(
     const offset_t * stride_out,
     const offset_t * stride_hes,
     const offset_t * stride_wgt,
-          int        stream = 0)
+          intptr_t   stream = 0)
 {
     const offset_t ndim = nbatch + 1;
     offset_t * d_size = nullptr, * d_so = nullptr, * d_sh = nullptr, * d_sw = nullptr;
@@ -513,7 +513,7 @@ void sym_invert(
     const offset_t * size,
     const offset_t * stride_out,
     const offset_t * stride_hes,
-          int        stream = 0)
+          intptr_t   stream = 0)
 {
     const offset_t ndim = nbatch + 1;
     offset_t * d_size = nullptr, * d_so = nullptr, * d_sh = nullptr;
@@ -540,7 +540,7 @@ void sym_invert_(
           scalar_t * hes,
     const offset_t * size,
     const offset_t * stride,
-          int        stream = 0)
+          intptr_t   stream = 0)
 {
     const offset_t ndim = nbatch + 1;
     offset_t * d_size = nullptr, * d_st = nullptr;
